Add Selection Form/Page for QR Code Generation
Adds two new components, Select and SelectForm. Select represents the screen on which the user can select the information they want to use to generate a code, and the SelectForm is the actual form that is submitted. Also modified App.js to add a route for testing purposes. Currently the buttons do not have functionality, but the data is mocked from Mockaroo.