Add linter config, select all button, api data
Adds packages required for linting, routing, and GETting data from an api. Mock data is now retrieved by the Select Info page to be displayed in a checkbox list. A new Select All button is added. Checked-ness of boxes is now properly tracked. Currently there is some backup data in SelectForm.js for when Mockaroo has cut me off for the day, but should be removed later. Still need CSS.